+++ spec/fabricators/user_activity_fabricator.rb
@@ -16,6 +16,9 @@ Fabricator(:user_activity) do
   pr_count 3
   calories 870.2
   map { Fabricate.build(:map) }
+  before_validation do
+    self.team ||= user.team
+  end
   after_create do
     map&.save!
   end